vm_info.rs renamt to simulator_context
