+++ src/simulator_context.rs
