+++ src/spawn.rs
@@ -2,7 +2,7 @@ use crate::{
     get_cur_tx, get_cur_tx_mut, get_cur_vm,
     global_data::{GlobalData, VmID},
     utils,
-    vm_info::{Fd, TxContext, VMInfo},
+    simulator_context::{Fd, TxContext, VMInfo},
 };
 use std::os::raw::{c_int, c_void};
